Make ComposableArchitecture group.
